+++ src/heap/scavenger-inl.h
@@ -461,11 +461,11 @@ class ScavengeVisitor final : public NewSpaceVisitor<ScavengeVisitor> {
 
   V8_INLINE void VisitPointers(Tagged<HeapObject> host, MaybeObjectSlot start,
                                MaybeObjectSlot end) final;
-  V8_INLINE int VisitEphemeronHashTable(Tagged<Map> map,
-                                        Tagged<EphemeronHashTable> object);
-  V8_INLINE int VisitJSArrayBuffer(Tagged<Map> map,
-                                   Tagged<JSArrayBuffer> object);
-  V8_INLINE int VisitJSApiObject(Tagged<Map> map, Tagged<JSObject> object);
+  V8_INLINE size_t VisitEphemeronHashTable(Tagged<Map> map,
+                                           Tagged<EphemeronHashTable> object);
+  V8_INLINE size_t VisitJSArrayBuffer(Tagged<Map> map,
+                                      Tagged<JSArrayBuffer> object);
+  V8_INLINE size_t VisitJSApiObject(Tagged<Map> map, Tagged<JSObject> object);
   V8_INLINE void VisitExternalPointer(Tagged<HeapObject> host,
                                       ExternalPointerSlot slot);
 
@@ -523,16 +523,16 @@ void ScavengeVisitor::VisitPointersImpl(Tagged<HeapObject> host, TSlot start,
   }
 }
 
-int ScavengeVisitor::VisitJSArrayBuffer(Tagged<Map> map,
-                                        Tagged<JSArrayBuffer> object) {
+size_t ScavengeVisitor::VisitJSArrayBuffer(Tagged<Map> map,
+                                           Tagged<JSArrayBuffer> object) {
   object->YoungMarkExtension();
   int size = JSArrayBuffer::BodyDescriptor::SizeOf(map, object);
   JSArrayBuffer::BodyDescriptor::IterateBody(map, object, size, this);
   return size;
 }
 
-int ScavengeVisitor::VisitJSApiObject(Tagged<Map> map,
-                                      Tagged<JSObject> object) {
+size_t ScavengeVisitor::VisitJSApiObject(Tagged<Map> map,
+                                         Tagged<JSObject> object) {
   int size = JSAPIObjectWithEmbedderSlots::BodyDescriptor::SizeOf(map, object);
   JSAPIObjectWithEmbedderSlots::BodyDescriptor::IterateBody(map, object, size,
                                                             this);
@@ -565,8 +565,8 @@ void ScavengeVisitor::VisitExternalPointer(Tagged<HeapObject> host,
 #endif  // V8_COMPRESS_POINTERS
 }
 
-int ScavengeVisitor::VisitEphemeronHashTable(Tagged<Map> map,
-                                             Tagged<EphemeronHashTable> table) {
+size_t ScavengeVisitor::VisitEphemeronHashTable(
+    Tagged<Map> map, Tagged<EphemeronHashTable> table) {
   // Register table with the scavenger, so it can take care of the weak keys
   // later. This allows to only iterate the tables' values, which are treated
   // as strong independently of whether the key is live.
